Rollover recap log only when it is written.
Previousely, this log was rolled over at the start
of every ansible-pull run. Leading to a state where
no recap log file was present.

When getting status information with ants -s, this
would result in a "failed" state being returned.
(No log was found, so we assumed failed.)

This commit fixes that behaviour by making sure
the log is only rolled over imediately befor it
is written.