Fixing the byte/string problem
In python 3 bytestrings and strings are handled separately.
Therefore, the output of a subprocess, which is a byte string,
must be explicitly decoded to a utf-8 string to work with the logger.

Additionally the subprocess has to stop if b"" is the output.

This has worked before in python 2 because "" == b"". It does now still
work in python2 with the change added.