Adding a first draft
Inside the directory `mac_os_relocatable` a new Makefile exists now which allows it to create a ants pkg with an included relocatable python.

This works for python2 and 3. If the package should be build with python2 this has to be explicitely specified as `sudo make pkg PYTHON_VERSION=2.7.15`.

The LaunchDaemons and the post install script were adapted and a requirements file was added which is used for the creation of the relocatable python.