Possible fix for inventory script hashbang problem
At the moment if ANTS is installed via pip, the hashbang lines are not
adapted because they are not recognized as scripts. Therefore they are
now added in the setup.py as scripts. This should make sure the hashbang
line points to the correct python interpreter when installing via pip.